update main.js
to summarize, if the user tries to submit an empty to-do item, the code prevents the data from being saved to local storage, provides an alert to inform the user, and gives them instructions on what to do next (enter a valid to-do item). If the input is not empty, the to-do item is added to the list as usual.