updates to styles and index
fixed issue with footer not sticking to btn.
It was bcs main content did not fill the page properly which made the footer act strange.
Gave the main content a class/section and gave it property 100vh height to always fill the content. This fixed the issue with the footer